Code cleaning : remove Fic_Attr , replace by Attr_Def_SQL

+++ include/class/card_attribut_mtable.class.php
@@ -148,13 +148,11 @@ class Card_Attribut_MTable extends Manage_Table_SQL
     
     function check()
     {
-        $nb_error=0;
         $object_sql=$this->get_table();
 
         if ($object_sql->get("ad_id")<9000 && $object_sql->get("ad_id")!=-1)
         {
             $this->set_error("ad_id",_("Bloqué"));
-            return false;
         }
         
         // Duplicate
@@ -168,30 +166,70 @@ class Card_Attribut_MTable extends Manage_Table_SQL
         if ($count_duplicate>0)
         {
             $this->set_error("ad_text", _("Doublon"));
-            $nb_error++;
         }
+        // type
+         if (in_array($object_sql->get("ad_type"),
+                 array('date', 'text', 'numeric', 'zone', 'poste', 'card', 'select','check'))==false)
+         {
+             $this->set_error("ad_type", _("Type invalide"));
+         }
 
         // Name empty
         if (trim($object_sql->ad_text)=="")
         {
             $this->set_error("ad_text", _("Description ne peut pas être vide"));
-            $nb_error++;
         }
         // select protect 
+        if ( $object_sql->ad_size=="") {
+            $object_sql->ad_size=22;
+        }
+        if ($object_sql->ad_type=='numeric')
+        {
+            $object_sql->ad_extra=(trim($object_sql->ad_extra)=='')?'2':$object_sql->ad_extra;
+            if (isNumber($object_sql->ad_extra)==0)
+            {
+                 $this->set_error("ad_text",_("La précision doit être un chiffre"));
+            }
+        }
+        
+        if ($object_sql->ad_type=='select')
+        {
+            if (trim($object_sql->ad_extra)=="") {
+                
+                $this->set_error("ad_extra",_("La requête SQL est vide "));
+            }
+            if (preg_match('/^\h*select/i', $object_sql->ad_extra)==0)
+            {
+               $this->set_error("ad_extra",_("La requête SQL doit commencer par SELECT "));
+            }
 
-
-        if ($nb_error>0)
+        }
+        if ($this->count_error()>0) {
             return false;
+        }
         return true;
     }
 
     function delete() {
          $object_sql=$this->get_table();
+         
+         if ( $object_sql->getp("ad_id") < 9000 ) {
+             throw new Exception (_("Effacement bloqué"));
+         }
+         
          $db=$object_sql->cn;
          try {
              $db->start();
-              $object_sql=$this->get_table()->delete();
-             $db->commit();
+             $sql=$db->exec_sql("delete from fiche_detail  where ad_id=$1 ", array($object_sql->ad_id));
+
+            $sql="delete from jnt_fic_attr where ad_id=$1";
+            $res=$db->exec_sql($sql, array($object_sql->ad_id));
+
+            $sql="delete from attr_def where ad_id=$1";
+            $res=$db->exec_sql($sql, array($object_sql->ad_id));
+            
+            $object_sql=$this->get_table()->delete();
+            $db->commit();
          } catch (Exception $ex) {
              $db->rollback();
             throw new Exception (_("Effacement bloqué : attribut utilisé"));
